+++ crates/events/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "starbase_events"
+version = "0.1.0"
+edition = "2021"
+license = "MIT"
+description = "Macros for the starbase framework."
+repository = "https://github.com/moonrepo/starbase"
+
+[dependencies]
+async-trait = "0.1.68"
+miette = "5.7.0"
+tokio = { version = "1.27.0" }
