+++ crates/events/README.md
@@ -1,0 +1,6 @@
+# starbase_events
+
+![Crates.io](https://img.shields.io/crates/v/starbase_events)
+![Crates.io](https://img.shields.io/crates/d/starbase_events)
+
+TODO
