+++ crates/events/src/lib.rs
@@ -1,0 +1,7 @@
+mod emitter;
+mod event;
+mod subscriber;
+
+pub use emitter::*;
+pub use event::*;
+pub use subscriber::*;
